Guide Page > Main Summary Area 추가
