+++ src/guide/GuideApp.vue
@@ -12,15 +12,9 @@
             </div>
         </div>
 
-        <div class="container">
+        <div class="center">
             <cmp-nav v-on:getVueFile="getVueFile"></cmp-nav>
-            <div class="right">
-                <cmp-content :fileList="vueFileList"></cmp-content>
-                <cmp-result ref="cmpResult" :fileList="vueFileList"></cmp-result>
-            </div>
-
-            <!--<cmp-content ref="cmpContent" @emitContent="emitContent"></cmp-content>-->
-            <!--<div><router-view name="Result"></router-view></div>-->
+            <cmp-content class="content-area" :fileList="vueFileList"></cmp-content>
         </div>
     </div>
 </template>
@@ -37,15 +31,6 @@
             'cmpContent': Content,
             'cmpResult': Result
         },
-        computed:{
-          initText: function () {
-            return {
-              value : `EVUI는 웹페이지의 핵심 구성요소인 Grid/Chart 컴포넌트를 제공하는 UI 프레임워크입니다.
-                       EVUI Grid와 Chart는 HTML/CSS/JS 및 SVG로 구현되어 있어 다양한 환경에 적용이 가능하며,
-                       Vue.JS를 기반으로 구현되어 대량의 데이터를 고속으로 처리합니다.`
-            }
-          }
-        },
         data: function () {
             return {
                 isLoading: false,
@@ -54,14 +39,6 @@
             }
         },
         methods: {
-            emitContent: function (id) {
-                var contentView;
-
-                if ( id == 'updateResult' ) {
-                    contentView = this.$refs.contentView;
-                    this.$refs.cmpResult.update( contentView.getContentName(), contentView .getValue() );
-                }
-            },
             getVueFile: function (path) {
                 const baseURI = '../../static/';
                 var vm = this;
@@ -72,46 +49,45 @@
                 }
 
                 this.$http.get(`${baseURI}${fileName}.vue`)
-                        .then((result) => {
+                .then((result) => {
                     let tmpObj = vm.codeParser(result.data);
-                    if(tmpObj){
-                        vm.$set(vm.vueFileList, fileName, tmpObj);
-                    }
-                }, (err) => {});
-            },
-            codeParser: function(data = null , ...rest){
-                let ix, ixLen;
-                let startTag, endTag;
-                let startIndex, endIndex;
-                let keyList;
-
-                let obj = {
-                    template: '',
-                    style: '',
-                    script: ''
-                };
-
-                if(!data){
-                    return data;
+                if(tmpObj){
+                    vm.$set(vm.vueFileList, fileName, tmpObj);
                 }
+            }, (err) => {});
+    },
+    codeParser: function(data = null , ...rest){
+        let ix, ixLen;
+        let startTag, endTag;
+        let startIndex, endIndex;
+        let keyList;
 
-                keyList = Object.keys(obj);
+        let obj = {
+            template: '',
+            style: '',
+            script: ''
+        };
 
-                for(ix = 0, ixLen = keyList.length; ix < ixLen; ix++){
-                    startTag = `<${keyList[ix]}>`;
-                    endTag = `</${keyList[ix]}>`;
-                    keyList[ix] === 'style' ? startIndex = data.lastIndexOf(startTag) : startIndex = data.indexOf(startTag);
-                    endIndex = data.lastIndexOf(endTag);
-                    obj[keyList[ix]] = data.substring(startIndex + startTag.length, endIndex).trim();
-                }
-
-                return obj;
-            }
-        },
-        mounted: function(){
-            this.getVueFile('ContentA');
+        if(!data){
+            return data;
         }
 
+        keyList = Object.keys(obj);
+
+        for(ix = 0, ixLen = keyList.length; ix < ixLen; ix++){
+            startTag = `<${keyList[ix]}>`;
+            endTag = `</${keyList[ix]}>`;
+            keyList[ix] === 'style' ? startIndex = data.lastIndexOf(startTag) : startIndex = data.indexOf(startTag);
+            endIndex = data.lastIndexOf(endTag);
+            obj[keyList[ix]] = data.substring(startIndex + startTag.length, endIndex).trim();
+        }
+
+        return obj;
+    }
+    },
+    mounted: function(){
+        this.getVueFile('ContentA');
+    }
     }
 </script>
 
@@ -154,7 +130,7 @@
         font-size: 20px;
     }
 
-    .guide-app .container {
+    .center {
         position: absolute;
         top: 60px;
         left: 0px;
@@ -162,7 +138,8 @@
         bottom: 0px;
         height: auto !important;
     }
-    .guide-app .container .right {
+
+    .center .content-area {
         padding: 0 !important;
         position: absolute;
         top:0px;
@@ -171,25 +148,7 @@
         bottom:0px;
     }
 
-
-    button {
-      height: 29px;
-      width: 100px;
-      border-radius: 6px;
-      color: #fff;
-      background-image: -webkit-linear-gradient(top, #55d1ff 0%, #00b4f7 100%);
-      border: 1px solid #4d899d;
-    }
-    button:hover {
-      border-radius: 6px;
-      color: #fff;
-      background-image: -webkit-linear-gradient(top, #40abf5 0%, #0091f7 100%);
-      border: 1px solid #4d899d;
-      text-shadow: 0px 1px 0px rgba(16,126,173,.9);
-      cursor: pointer;
-    }
-
     p {
-      font-size: 15px;
+        font-size: 15px;
     }
 </style>
